requirements.txt
Django
Pillow
django-crispy-forms
python-dotenv
django-sendgrid-v5
boto3
django-storages